requirements.txt
flask
flask-cors
pyairtable
google-adk
dotenv
pandas
google-cloud-bigquery
google-cloud-firestore
google-cloud-storage
google-cloud-aiplatform
